CHANGED LOOP
Added the loop that changes how the game looks. I wanted to make it harder for the player to find the ball. The more balls the merrier. The for loop, much like the static for loop, draws multiple balls on the screen only one of which is the real one.