Changed Loop Name
Changed the shape of the ball itself. I find it more pleasing when the ball is an actual ball and not a square. Also changed the same thing in the loop, so the others balls aren't squares anymore which I find makes it even harder to find the actual ball.