Added New Function
Added the new function with checks where the ball is on the screen and depending on where it is, the paddle adapts and changes width. If the ball is on the left of the screen, the paddle will get significantly smaller while if its on the right side of the screen, the paddle will change size again.