Think I Made Something
I think I accidentally got the two creatures to interact with one another. Haven't added the comments yet, but basically what I added checks whether the griddies' x and y location is the same as the grunties' x and y location and if it is, the griddies become infected and change to a pink colour.