Changes to ALL Pong Games
Added a reset of the left paddle and right paddle score. For some reason whenever I would play one of the games and return to the menu, and then again return to the same game, the score would be the same. This just simply resets the score to 0 whenever the user decides to go back to the menu.